SI-6766 Makes the -Pcontinuations:enable flag a project specific preference
Make the continuations plugin flag specific to the continuations library.
Remove the documentation about adding the continuation plugin from
the eclipse README.md.

One consequence of this change is that if a programmer has prefs
for the continuations eclipse project then there will be a merge conflict.
That's unavoidable without having a more sophisticated way  to build
the eclipse projects.